Reduce less often and only after the data are outputed
Reduce only after BLOCK*5 packets are stored. Previously the first reduction resulted in reduction
after each received BLOCK but this is not necessary after the reduce_packets() was fixed. Also in
rare cases where the number of packets is not reduced enough, one BLOCK of packets is removed in
oreder to reduce the number of reduce_packets() calls.