Big ugly commit - rewrite to C++
Packet and computer database rewritten to C++, the functionality should be very similar to C-version
but clock skew comparision doesn't work at the moment.

There are new clases for:
* packet time info
* computer info
* all computers with program settings

STL used instead of own lists. It should be clear now which functions work with all computers and
which with only one. Need for global variables reduced.

Some parts of code are commented out. It will be cleaned later.