log_reader: Set maximal Time limit
The goal is to process the log file as is. So the output should be as
close to the input as possible. Hence we do not remove known clock skew
computation only because there were a gap in their clock skew
computation. The gap was probably caused by removing unnecessary points
under the convex hull.